worker/requirements.txt
google-cloud==0.27
requests==2.11.1
click==6.6
healthcheck==1.3.1
Flask==0.11
Pillow==3.3.0